+++ src/config.rs
@@ -18,6 +18,9 @@ pub struct Config {
     /// Whether to show receipts (sent, delivered, read) information next to your user name in UI
     #[serde(default = "default_true")]
     pub show_receipts: bool,
+    /// Whether to show system notifications on incoming messages
+    #[serde(default = "default_true")]
+    pub notifications: bool,
     /// User configuration
     pub user: User,
 }
@@ -39,6 +42,7 @@ impl Config {
             signal_db_path: default_signal_db_path(),
             first_name_only: false,
             show_receipts: true,
+            notifications: true,
         }
     }